requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
playwright==1.40.0
watchdog==3.0.0
psutil==5.9.6
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0